projeto_django/requirementes.dev.txt
django<2.1
psycopg2<2.8
django-debug-toolbar